# skills/bkd/references/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BKD API",
    "description": "Kanban board for managing AI coding agents. Issues are assigned to CLI-based AI engines (Claude Code, Codex, Gemini CLI) that execute autonomously.",
    "version": "dev",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Default"
    }
  ],
...
  "components": {
    "schemas": {
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "alias": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "directory": {
            "type": "string"
          },
          "repositoryUrl": {
            "type": "string"
          },
          "systemPrompt": {
            "type": "string"
          },
          "envVars": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "sortOrder": {
            "type": "string"
          },
          "isArchived": {
            "type": "boolean"
          },
          "isGitRepo": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "alias",
          "name",
          "sortOrder",
          "isArchived",
          "createdAt",
          "updatedAt"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
      "CreateProject": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "alias": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "pattern": "^[a-z0-9]+$"
          },
          "description": {
            "type": "string",
            "maxLength": 5000
          },
          "directory": {
            "type": "string",
            "maxLength": 1000
          },
          "repositoryUrl": {
            "anyOf": [
              {
                "type": "string",
                "format": "uri"
              },
              {
                "type": "string",
                "enum": [
                  ""
                ]
              }
            ]
          },
          "systemPrompt": {
            "type": "string",
            "maxLength": 32768
          },
          "envVars": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "maxLength": 10000
            }
          }
        },
        "required": [
          "name"
        ]
      },
...
      "Issue": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "statusId": {
            "type": "string",
            "enum": [
              "todo",
              "working",
              "review",
              "done"
            ]
          },
          "issueNumber": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "tags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "sortOrder": {
            "type": "string"
          },
          "useWorktree": {
            "type": "boolean"
          },
          "isPinned": {
            "type": "boolean"
          },
          "keepAlive": {
            "type": "boolean"
          },
          "engineType": {
            "type": [
              "string",
              "null"
            ]
          },
          "sessionStatus": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "pending",
              "running",
              "completed",
              "failed",
              "cancelled",
              null
            ]
          },
          "prompt": {
            "type": [
              "string",
              "null"
            ]
          },
          "externalSessionId": {
            "type": [
              "string",
              "null"
            ]
          },
          "model": {
            "type": [
              "string",
              "null"
            ]
          },
          "statusUpdatedAt": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "statusId",
          "issueNumber",
          "title",
          "tags",
          "sortOrder",
          "useWorktree",
          "isPinned",
          "keepAlive",
          "engineType",
          "sessionStatus",
          "prompt",
          "externalSessionId",
          "model",
          "statusUpdatedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateIssue": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 50
            },
            "maxItems": 10
          },
          "statusId": {
            "type": "string",
            "enum": [
              "todo",
              "working",
              "review",
              "done"
            ]
          },
          "useWorktree": {
            "type": "boolean"
          },
          "keepAlive": {
            "type": "boolean"
          },
          "engineType": {
            "type": "string",
            "pattern": "^(claude-code|codex|acp(:.+)?)$",
            "description": "claude-code | codex | acp | acp:<agent>:<model>"
          },
          "model": {
            "type": "string",
            "pattern": "^[\\w./:\\-[\\]]{1,160}$"
          },
          "permissionMode": {
            "type": "string",
            "enum": [
              "auto",
              "supervised",
              "plan"
            ]
          }
        },
        "required": [
          "title",
          "statusId"
        ]
      },
...
      "WorktreeEntry": {
        "type": "object",
        "properties": {
          "issueId": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "branch": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "issueId",
          "path",
          "branch"
        ]
      },
...
    },
...
  },
  "paths": {
    "/api/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "List projects",
        "operationId": "listProjects",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "archived",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Project list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Project"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Create project",
        "operationId": "createProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProject"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created project",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/Project"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Directory already in use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}/issues": {
      "get": {
        "tags": [
          "Issues"
        ],
        "summary": "List issues in project",
        "operationId": "listIssues",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "projectId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Issue list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Issue"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Issues"
        ],
        "summary": "Create issue",
        "operationId": "createIssue",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "projectId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIssue"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created issue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/Issue"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "202": {
            "description": "Created and executing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/Issue"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}/worktrees": {
      "get": {
        "tags": [
          "Worktrees"
        ],
        "summary": "List worktrees for project",
        "operationId": "listWorktrees",
        "responses": {
          "200": {
            "description": "Worktree list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/WorktreeEntry"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}